Fix options bug; add ability to skip missing migrate sources
The `options:` setting for `migrate` actions was reading an erroneous truthiness to test for the existence of a setting.
Fixed this so opt settings other than `inclusive: false` actually work.

Added the ability to set an option `missing:` with either `warn` or `raise` as values.
If `missing: warn`, the operation skips the migrate step and logs a warning to console.
If `missing: raise` (the default, w/ backward compatibility) the operation throws an error and exits.

Added example and explanation for new `missing:` setting to README.